requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.3.0
attrs==23.2.0
av==9.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dnspython==2.6.1
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
idna==3.7
ifaddr==0.2.0
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-option-menu==0.3.13
streamlit-webrtc==0.47.7
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
thop==0.1.1.post2209072238
tkvideoplayer==2.8
toml==0.10.2
toolz==0.12.1
torch==2.3.1 
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.34
ultralytics-thop==2.0.0
urllib3==2.2.1
watchdog==4.0.1
